requirements.txt
django
djangorestframework
pygments
# Heroku reqs
django-heroku
gunicorn
whitenoise
#IPython
ipython
